+++ src/components/qam/common/ShelfListLabel.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+import { icons } from '../icons'
+
+export function ShelfListLabel({ shelf }: { shelf: any }) {
+  return (
+    <div className={`deck-shelves-label-cont ${shelf.hidden ? 'deck-shelves-hidden' : ''}`}>
+      <span className='deck-shelves-hidden-icon'>{shelf.hidden ? icons.eyeClosed : icons.eyeOpen}</span>
+      <span className='deck-shelves-label-text'>{shelf.title}</span>
+    </div>
+  )
+}
