+++ src/components/qam/modals/DeleteConfirmModal.tsx
@@ -1,0 +1,27 @@
+import React from 'react'
+import { ConfirmModal } from '@decky/ui'
+import type { SettingsController } from '../../features/settings/controller'
+import type { Shelf } from '../../types'
+
+export function DeleteConfirmModal({ closeModal, controller, shelf }: { closeModal?: () => void; controller: SettingsController; shelf: Shelf }) {
+  const { t, actions } = controller
+
+  return (
+    <ConfirmModal
+      strTitle={t('deleteShelf')}
+      strDescription={shelf.title}
+      strOKButtonText={t('deleteShelf')}
+      strCancelButtonText={t('cancel')}
+      bDestructiveWarning
+      onCancel={closeModal}
+      onEscKeypress={closeModal}
+      onOK={() => {
+        closeModal?.();
+        (async () => {
+          const ok = await actions.removeShelf(shelf.id);
+          // lightweight log kept in parent file
+        })();
+      }}
+    />
+  )
+}
