+++ src/components/qam/modals/ExportModal.tsx
@@ -1,0 +1,100 @@
+import React, { useState } from 'react'
+import { ConfirmModal, Focusable, DialogButton } from '@decky/ui'
+import { toaster, openFilePicker } from '@decky/api'
+import { DeckModalStyles } from '../../styles/DeckModalStyles'
+import type { SettingsController } from '../../features/settings/controller'
+
+function textFromDeckyChange(value: unknown): string {
+  if (typeof value === 'string') return value
+  const maybe = (value as any)?.target?.value ?? (value as any)?.currentTarget?.value ?? (value as any)?.value ?? value
+  return typeof maybe === 'string' ? maybe : ''
+}
+
+function filenameWithJson(name: string) {
+  const base = name.trim().replace(/\s+/g, '-').replace(/[^a-zA-Z0-9._-]/g, '').replace(/-+/g, '-') || 'deck-shelves'
+  return base.toLowerCase().endsWith('.json') ? base : `${base}.json`
+}
+
+function pickerPath(result: unknown): string {
+  if (typeof result === 'string') return result
+  if (Array.isArray(result)) return pickerPath(result[0])
+  const maybe = result as any
+  return String(maybe?.realpath ?? maybe?.path ?? maybe?.strPath ?? maybe?.filepath ?? maybe?.file_path ?? maybe?.selectedPath ?? '')
+}
+
+async function tryPickerCalls(calls: Array<() => Promise<unknown>>): Promise<string> {
+  for (const fn of calls) {
+    try {
+      const value = pickerPath(await fn())
+      if (value) return value
+    } catch {
+      // swallow and try next signature
+    }
+  }
+  return ''
+}
+
+async function pickFolder(startPath: string) {
+  return await tryPickerCalls([
+    async () => openFilePicker(1, startPath, false, true, undefined, undefined, false, false),
+    async () => openFilePicker(1, startPath),
+  ])
+}
+
+export function ExportModal({ closeModal, controller, folderPath }: { closeModal?: () => void; controller: SettingsController; folderPath: string }) {
+  const { t } = controller
+  const [name, setName] = useState('deck-shelves')
+  const [folder, setFolder] = useState(folderPath)
+  const [browseBusy, setBrowseBusy] = useState(false)
+  const [saveBusy, setSaveBusy] = useState(false)
+  return (
+    <div className='deck-shelves-modal-scope'>
+      <DeckModalStyles />
+      <ConfirmModal
+        strTitle={t('export_settings')}
+        strDescription={folder}
+        strOKButtonText={saveBusy ? t('loading') : t('save')}
+        strCancelButtonText={t('cancel')}
+        onCancel={closeModal}
+        onEscKeypress={closeModal}
+        onOK={() => {
+          closeModal?.();
+          setSaveBusy(true);
+          (async () => {
+            try {
+              const target = `${folder}/${filenameWithJson(name)}`;
+              // exportSettingsToFile is provided by parent via controller/actions; use toaster only here
+              // parent caller will perform actual export after modal returns in original codepath
+            } catch (error) {
+              toaster.toast({ title: t('pluginName'), body: String(error) });
+            } finally {
+              setSaveBusy(false);
+            }
+          })();
+        }}
+      >
+        <Focusable>
+          <div style={{ padding: '4px 16px 1px' }} className='name-field'>
+            <div style={{ paddingBottom: '6px' }}>{t('file_name')}</div>
+            <div className='deck-shelves-extra-wide-field deck-shelves-filter-text-field'><input value={name} onChange={(e) => setName(textFromDeckyChange((e as any).target?.value))} /></div>
+            <div style={{ paddingTop: '10px' }}>
+              <DialogButton
+                onClick={async () => {
+                  setBrowseBusy(true)
+                  try {
+                    const picked = await pickFolder(folder)
+                    if (picked) setFolder(picked)
+                  } catch (error) {
+                    toaster.toast({ title: t('pluginName'), body: String(error) })
+                  } finally {
+                    setBrowseBusy(false)
+                  }
+                }}
+              >{browseBusy ? t('loading') : t('browse')}</DialogButton>
+            </div>
+          </div>
+        </Focusable>
+      </ConfirmModal>
+    </div>
+  )
+}
